requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.21.1
langchain==0.0.320
openai==1.3.7
pinecone-client==2.2.1
chromadb==0.3.29
spacy==3.6.0
pdfminer.six==20221105
python-multipart==0.0.6
streamlit==1.26.0
sqlalchemy==1.4.50
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.12
typing_extensions==4.5.0
